% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang.R
\name{ang}
\alias{ang}
\title{Calculate the angle between successive segments along a coastline}
\usage{
ang(line)
}
\arguments{
\item{line}{A way tagged with natural=coastline}
}
\value{
a vector of angles
}
\description{
Calculate the angle between successive segments along a coastline
}
\examples{
 # Download the coastline from the Río Negro province in Argentina
RN.coast <- getosmcoastline("Río Negro")

# Angles of the first way
ang(RN.coast[1,])
}
