% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkosmcoast.R
\name{checkosmcoast}
\alias{checkosmcoast}
\title{Analize a coast section}
\usage{
checkosmcoast(
  osmcoastline,
  thrangle = 30,
  thrlong = 500,
  pgs = TRUE,
  thrperc = 10
)
}
\arguments{
\item{osmcoastline}{Output of the function 'getosmcoastline' or a character vector with place name}

\item{thrangle}{A threshold angle in degrees}

\item{thrlong}{A threshold length in meters}

\item{pgs}{Logical. It must to check segments with PGS source?}

\item{thrperc}{Percentage of segments above which the coastline way needs revision}
}
\value{
An sf object
}
\description{
Analize a coast section
}
\examples{
# Download the coastline from the Río Negro province in Argentina
RN.coast <- getosmcoastline("Río Negro")

# Check three segments
RN.check <- checkosmcoast(RN.coast[3:5,])

# Map the selected segments
plot(RN.check)
}
