% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getosmcoastline.R
\name{getosmcoastline}
\alias{getosmcoastline}
\title{Get the OSM Coastline}
\usage{
getosmcoastline(place_name)
}
\arguments{
\item{place_name}{The name of the administrative boundary from which
to extract the coastline.}
}
\value{
An object of class 'sf' with the selected coastline
}
\description{
This function gets from OpenStreetMat the line tagged with "natural=coastline"
in a selected country
}
\examples{
RN <- getosmcoastline("Río Negro")
}
