% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{long}
\alias{long}
\title{Calculate the length of successive segments along a coastline portion}
\usage{
long(line)
}
\arguments{
\item{line}{A way tagged with natural=coastline}
}
\value{
A vector of the segments length in meters
}
\description{
Calculate the length of successive segments along a coastline portion
}
\examples{
 # Download the coastline from the Río Negro province in Argentina
RN.coast <- getosmcoastline("Río Negro")

# Lengths of the first way
long(RN.coast[1,])
}
